Revert "router added, but simply calling the client index page"
This reverts commit 41a428c146875fcf3aea68662a57b9af2cfd3f14.
