+++ webpack.config.js
@@ -13,7 +13,7 @@ module.exports = {
       {
         test : /\.jsx?/,
         include : SRC_DIR,
-        loader : 'babel-loader',
+        loader : 'babel-loader',      
         query: {
           presets: ['react', 'es2015']
        }
